Add attributes recognition for mediacapture
When capture attribute is present and one of accept types
is one of "video/*", "image/*" or "audio/*", file should be
captured from camera. In other cases, file selector
should be launched.

In this solution I assume that if there is at least one of
accept types mentioned above (and there may be many of them),
mediacapture will be launched.

Issue Id : CBWEBCORE-315, CBMM-129

Change-Id: Ib94dc90a09affdc657fff5c385d543e43e6d22b7