Enable touch events on desktop builds, and refactor EWebView::GetSnapShotForRect method.
In current implementation touch events are actually implemented on top
of EFL mouse events. They are only enabled on tizen builds however. It's
beneficial however to be able to also test them on desktop. Instead of
enabling touch events depending on the OS_TIZEN conditional make the
code rely on --touch-events command line flag. This way the touch events
will be by default enabled on the target builds (The flag is added in
command_line_efl.cc for tizen builds) and on desktop the flag can be
easily enabled by hand, ex:
$ efl_webview_app --touch-events

Also it refactors EWebView::GetSnapShotForRect.

Change-Id: I6b94ff0cb77fb1b181e388e03cd1de18fc0b4bf0

Conflicts:

	impl/eweb_view.cc