Improve text selection handling
This patch fixes several selection handle drawing issues caused by the
fact that SelectionBoxEfl was only fully updated from after mouse up event
handler was invoked. If a user manipulated the handles in a way that the
handles needed to be updated several times during one selection the end
results were not always matching the real state of the selection.

Long term, it'd be good if the selection state could be storred only in
one class. As it is right now the actual state of the selection resides
in the SelectionBoxEfl, while the visible state of the handles is in
SelectionHandleEfl. Those states are 'synchronized' in
SelectionControllerEfl. Such approach is very error prone as it's not
always obvious which state is valid at any given point in time.

Patch also hide context menu during magnifier is showing.

Issue-Id: CBBROWSER-156

Change-Id: I8c13e6c2734b294b5e912ce3a9fa78899da7e6de

Conflicts:
	impl/browser/renderer_host/render_widget_host_view_efl.cc
	impl/browser/renderer_host/render_widget_host_view_efl.h
	impl/selection_controller_efl.cc
	impl/selection_controller_efl.h