[M34-Merge] ewk_view_url_request_set should load URL along with the given header and body of the HTTP request
Implemented API for POST and GET methods. HEAD, PUT and DELETE cannot
be easily implemented in chromium and, from what we know, are not used
by any application using this API. This function implementation passes
the Tizen Compatibility Tests.
Null pointers were passed to std::string constructor and raised an exception

This is a combination of two commits:
http://165.213.202.130:8080/#/c/71637/ by Piotr An Nguyen and
http://165.213.202.130:8080/#/c/71717/ by Kamil Klimek

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=8268
Reviewed by: Antonio Gomes, Kamil Klimek

Change-Id: Ib6a28c86699a13653af30ef613f7af79089614ff
Signed-off-by: Michał Pakuła vel Rutka <m.pakula@samsung.com>