Add generic support for EFL toolkit.
This is similar to TOOLKIT_VIEWS. Basically this adds a gyp flag and a
preprocessor define that can be used to control EFL specific code parts.
It's supposed to supplement building_for_tizen_* flags. Contrary to
those however it's supposed to enable non tizen specific code that can
be built on any platform with EFL libraries avaliable (So far only
desktop).

Change-Id: I96e08bdd8261faa69a40990d37fa3c505e848e45

Conflicts:

	src/build/common.gypi