Remove OnAccessibilityEvents method declaration from RenderWidgetHostViewEfl.
This has been moved in chromium from RenderWidgetHost to RenderFrameHost.

Commit message (of the removal):

"
Migrate accessibility from RenderView to RenderFrame and RVH to RFH.

In preparation for out-of-process iframes, move almost all accessibility
functionality from RenderView to RenderFrame, and from RVHI and RWHI
to RFHI.

In addition, the accessibility mode is moved to WebContents, and changing
the mode triggers changing it on all frames.

BUG=368298
R=jam@chromium.org, kenrb@chromium.org, nick@chromium.org

Review URL: https://codereview.chromium.org/273423004
"