ccache dir moved outside source directory in gbs builds.
Previously ccache directory for gbs was placed in
chromium-efl directory. It was reusable because by default all builds
were made with --incremental flag so chromium-efl repo dir
was mounted into gbs build root. Without -incremental flag
it is not possible anymore. However we can use /.ccache directory inside
gbs sysroot for that. It is not removed even when different package is
built or -C flag is passed.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=5263

Change-Id: Iab7481f38916f19877eeed0dde96fd11e45fb365
Signed-off-by: Tomasz Olszak <t.olszak@samsung.com>