[Tizen] Re-factor WebContents usage by EWK API implementation
WebContents is the main object of the content API from user
perspective. The object represents a web widget with extra
logic used to manipulate it's state (LoadUrl, go back, stop
loading page, etc). A lot of this logic maps directly to what
EWebView class is supposed to do. It's can be easily said that
EWebView is a wrapper aroud WebContents.
In the current codebase the actual ownership of WebContents is
rather strange. It's not held directly by it's main user, but by
WebContentsDelegateEfl. As a direct result we:
1. Do lot of unecessary web_contents_delegate_->web_contents()
   getter calls.
2. We often pass WebContentsDelegate to objects that only need
   WebContents
3. Have bunch of unnecessary WebContents NULL checks. The
   classes which make them could not even be instantiated if
   WebContents did not exist.
This patch moves the ownership of WebContsnts to EWebView.
It also changes, wherever possible, WebContents pointers to
references.

Original patch: http://165.213.202.130:8080/#/c/69163/3
Original author: Piotr Tworek <p.tworek@samsung.com>
Reviewed by:  Antonio Gomes

Change-Id: I507f270f2fa62ca70a4d24972359c8be5b5c3175
Signed-off-by: zhaoze <zhaoze.zhou@partner.samsung.com>