[Tizen][Build] Disable component build automagically
I regularly run into this. For open source developement the component
build is a great feature so it should be ok to keep it to be on by default.
This change adds a tiny bit of logic to the build system to disable it
automatically as it is not compatible with chromium-efl.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=6110

Change-Id: Ifcbbee3d8763f299752584b804108f41bac67b41
Signed-off-by: Balazs Kelemen <b.kelemen@samsung.com>