[Tizen][Build] Clean up dependency handling
Currently the depedency handling is done in a rather obscure way.
We add the dependencies needed by chromium-efl as target_defaults via
chromium-ef.gypi. This means that each and every target will use efl and other
include paths, link flags and the like. Obviously internal chromium modules
do not need this. One practical negative side-effect of this is that we execute
pkg-config thousands of times for a build.

This change refactors this by moving the dependencies to a new gyp file that
use targets that define the dependencies much like system.gyp does it in
chroimium. Now only the top level chromium-efl target will use the dependencies
resulting in simpler and faster build.

Tested with desktop and mobile build.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=6110

Change-Id: I4afd4576c8eed022c648be274a50ac3c3b3cd07e
Signed-off-by: Balazs Kelemen <b.kelemen@samsung.com>