Switch to Tizen 2.3
Minimal changes to be able to build with Tizen 2.3. Some features have been
disabled because of API changes. Fixing them is TODO.

The webrtc patch had to be resurrected to fix debug build (partially).

I switched to patching approach instead of copying patched files because
I believe it is more flexible and it shows if the patch is out-of-date more
directly.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=5477

Change-Id: I67606fca32e1af62552a084d3f850d556e2b848a
Signed-off-by: Balazs Kelemen <b.kelemen@samsung.com>