Fix gestures with m39
Chromium doesn't handle or want to see touch-still (i.e. EVAS_TOUCH_POINT_STILL)
events. In debug it leads to a NOTREACHED in aura code, in release it confuses
the gesture handler so it doesn't function correctly. It seems like we can
simply ignore these events.

Note that this doesn't solve the bringup of impl level touch features such as
selection or overscroll effect.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=5593

Change-Id: Ia8fe780d2ca507481641ee0ce9feb497783cfc7d
Signed-off-by: Balazs Kelemen <b.kelemen@samsung.com>