Build ewk with gyp instead of cmake
As of now, one executable is created for each unittest. There are about
300 of them. The gyp code to handle that would be cumbersome to write.
Also, in chromium upstream, all unittests of a specific subdirectory
create only one executable. So, we don't build the unittests for now. We
will first need to figure out what to do with them.

Also, modify spec files so that files which were previously in
chromium-ewk rpm are now included in chromium-efl

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=6096

Change-Id: I14e47fb42234122ce44a4ad6c169342dea1054d4
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>