IME is not shown when clipboard window is visible.
[Problem] When clipboard window is visible, there is
no possibility to show IME keyboard. It should be visible
after tap on input field.

[Solution] Added implementation of IsClipboardWindowOpened
method. It depends on ecore_x_e_illume_clipboard_state_get.
With this patch clipboard window is closed, when user tap
on not focused input field.
Also added closing clipboard window when webview is out of
focus.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=9716
This is a cherry-pick from: http://165.213.202.130:8080/#/c/70395/
Reviewed by: Piotr Grad, SeungSeop Park, commitbot

Change-Id: I42726064071255579504774345311cd954263244
Signed-off-by: Karol Furmaniak <k.furmaniak@samsung.com>