Enhance Tizen device development with faster deployment
Add rsync based fast push script and helpers to make deployment easy and fast.
The scripts upload libs and binaries as is (no rpm install) Envsetup script
provided for device.

This workflow has been chosen since it is suitable for both gbs and
cross-compile workflow. It is a developer-only setup, not in any sense
a replacement for rpm.

rpm based workflow can also benefit from rsync (using tizensync.sh directly).

Uploading libchromium-efl.so (stripped, release) with gbs takes ~34sec in my
environment. With this scripts it went down to ~8sec and the difference can
only be bigger when reuploading after build.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=6900
Reviewed by:  Viatcheslav Ostapenko, Antonio Gomes

Change-Id: I99c81a1f3b979e73a3d746cf5f4dd56b35380069
Signed-off-by: Balazs Kelemen <b.kelemen@samsung.com>