[M39] Fix tv build break
1) The character '\' is missed in 'build/build_tv.sh'.

2) The sensor library-level issue on tizen v2.1 and v2.2.1.

Error log is like below.
>> /usr/include/sensor/sensor.h:76:26: fatal error:
>> sensor_accel.h: No such file or directory

The '#include <sensor_accel.h>' is defined in 'sensor.h' that is
provided by platform rpm repository, and the 'sensor_accel.h' is
located in '/usr/include/sensor/'.

To use '#include <sensor_accel.h>' in 'sensor.h' file,
the path '/usr/include/sensor' needs to be defined in following pc file.
>> ./usr/lib/pkgconfig/capi-system-sensor.pc

However, on tizen v2.1 and v2.2.1,
the header path '/usr/include/sensor' is not included in pc file like below.
>> includedir=/usr/include/system

On the other hand, on tizen v2.3, it is properly included like below.
>> includedir=/usr/include/sensor

Therefore, the header path '/usr/include/sensor' needs to be defined in gyp file
for tizen v2.1 and v2.2.1.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=5676

Change-Id: Ie372f8e685eb7c829319b735ef41d4de2824b2e7
Signed-off-by: Youngsoo Choi <kenshin.choi@samsung.com>