Fixes crash when using virtual keyboard.
For whatever reason, when typing a text, once the second character has
been entered, EWebView::UpdateHitTestData is called. It crashes during
the copy of SkBitmap.

One way to prevent the crash is to avoid the copy of the
Hit_Test_Image_Buffer (and therefore the SkBitmap) in _Ewk_Hit_Test ipc
struct. Hit_Test_Image_Buffer is not used by the renderer.

In the long term, we should probably investigate more the underlying
nature of the crash. But for the moment, this patch allows the keyboard
to work without crashing.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=5456
Reviewed by:  Antonio Gomes

Change-Id: I9d7061e220a17d067f2a0e2b62c0c05b0af449a1
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>