Delete animator in SelectionMagnifierEfl::~SelectionMagnifierEfl
Right now, the animator set with ecore_animator_add is not deleted
in SelectionMagnifierEfl destructor. So, it can happen that the
animator callback is called after the magnifier has been deleted. This
results in a crash when trying to access its properties.

This patch fixes the problem by calling ecore_animator_del in
SelectionMagnifierEfl destructor.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=8053
Reviewed by: Balazs Kelemen, SeungSeop Park, Yi Shen

Change-Id: I91a9d9c86de54787fd37ae135b19ffd28c2eeaf6
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>