call ewk_set_arguments in unittests
Currently, the unittest binary crashes at the very beginning of it's
execution. This is because ewk_set_arguments is not called. And then,
the browser crashes in CommandLineEfl::GetDefaultPortParams at some
point later in the initialization.

This patch calls ewk_set_arguments before ewk_init

It also makes sure that CommandLineEfl::Init is always called before
CommandLineEfl::GetDefaultPortParams.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=6980
Reviewed by: Balazs Kelemen, SeungSeop Park

Change-Id: Idc8638de568b1921b8e0406562caf781dbc7620f
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>