use ExecuteJavaScriptForTests instead of ExecuteJavaScript
Full screen test fail because the full screen opening is blocked.
Actually, a fullscreen can only be opened if triggered by a gesture.

In chromium m34, executeJavascript used to act like if the script had
been executed by a gesture. But in recent chromium, there are two
different calls: executeJavascript and executeJavascriptForTests.

Since we don't want to change ewk api, we use executeJavascriptForTests
when a script is executed. This will mimick m34 behaviour

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=8558
Reviewed by: Antonio Gomes, Balazs Kelemen

Change-Id: I8e5025b826f7462aabdefa2a299f16222dba3f25
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>