ecore_x_init was not called
Currently, HAVE_ECORE_X is never set, and then, ecore_x_init is never
called from within ewk_init. Then, when we call ecore_x_display_get, the
display returned is null. This problem is not visible with
efl_webview_app because ecore_x_init is called indirectly (via a call to
elm_init). But for example, wrt does not call elm_init and triggers the
bug.

The proposed solution is to always consider that we're using ecore_x. It
is currently true for all our platforms.

This patch also fixes a logic error: ecore_imf_shutdown() should be
called whether or not ecore-x is available.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=9394
Reviewed by: Antonio Gomes, Piotr Tworek, arno renevier, commitbot

Change-Id: I5a41083f8e3cffe456fb638cd713d27a42d37620
Signed-off-by: Arnaud Renevier <a.renevier@samsung.com>