Enable "enable-view-port-meta" only on mobile
Some Web pages[1] are reported to have the following problems on displaying:
- Failed to use whole screen of TV
- Left aligned, rather than center aligned.
[1] naver.com, ebay.com, tumblr.com

This happens due to a bug on the implementation of viewport-meta tag
on Chromium. So the problems described above are also reproducible on
open source Chromium desktop version if #enable-viewport-meta at
chrome://flags is enabled.

Also, bug affects TV and desktop's only because Tizen mobile forces the
use of it's own viewport-meta width size regardless of the declared
viewport-meta width size by user.

Therefore, this patch fixes the bug by taking the following actions:

- Only allow AppendSwitch of kEnableViewportMeta for OS_TIZEN_MOBILE.

Bug : http://107.108.218.239/bugzilla/show_bug.cgi?id=9389
Reviewed by: Antonio Gomes, SeungSeop Park

Change-Id: I177a6cf97013217869fa50040e23d2a153991d40
Signed-off-by: dayoung.shin <dayoung.shin@samsung.com>