[M34-Merge] Attempt to gracefully handle renderer process crashes
The current chromium-efl tries to handle renderer process crashes by
invoking "webprocess,crashed" EFL callback on the EWebView. The callback
is expected to set the value of event_info to true and delete the
EWebView instance for which the crash did happen. After the callback is
invoked the crashed EWebView would be in an unusable state. Using it
would immediately cause the UI process to crash.

The problem with the current approach is only efl_webview_app actually
implements the callback. Other applications like mini_browser,
com.samsung.browser or com.samsung.email do not even register it. The
end result is webprocess crashes can actually bring down the entire app.
This defeats the purpose of chromium multi-process architecture.

This patch solves the problem by re-creating WebContents after the
renderer crash. The "webprocess,crashed" is now only used as a
notification mechanism. The embedder handling it is only responsible for
informing the user about the fact and possibly loading another page. If
it does not handle the callback the engine will show a simple error
page. The behaviour aligns chromium-efl with WebKitEfl behaviour.

Original patches:
1) "Attempt to gracefully handle renderer process crashes" by Piotr Tworek
   http://165.213.202.130:8080/#/c/69203
2) "Handle rendenrer crashes in efl_webview_app properly" by Piotr Tworek
   http://165.213.202.130:8080/#/c/69448

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=10194
Reviewed by: Janusz Majnert, Min-Soo Koo

Change-Id: I52407fbcd672f7726fb34bb83e911e6a784a149c
Signed-off-by: Piotr Szawdyński <p.szawdynski@samsung.com>