Hide content_shell URL bar when running in test mode
A couple of layout tests fail because they expect the web contents to
have the a size of 800x600 px. This value is given to the application
during init stage, but EFL shell uses it as the size of the application
window. It means the value is divided between the webview and URL bar.
The size of the URL bar depends on the current EFL theme so we can't
easilt resize the window to compensate for the difference.

Fortunately the URL bar is not actually needed by the tests themselves.
We can simply disable it, when running them. To detect such situation
--dump-render-tree flag is used. Chromium always specifies it when
running layout tests. In fact the flag is used to enable layout
test specific code paths in the upstream content_shell code.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=9684
Reviewed by: Antonio Gomes, Min-Soo Koo

Change-Id: I99f25b8ccd2f73677db94270ea4093f8d1d57dcf
Signed-off-by: Piotr Tworek <p.tworek@samsung.com>