Link chromium-efl statically when building content_shell_efl
Currently content_shell_efl application is built on top of both original
content_shell_lib component and chromium-efl library. This solution
allows us to avoid modification to the upstream files and gain access to
new code added for EFL port purposes. Unfortunately the solution has some
drawbacks. The one I was originally aware was that some source
files are compiled into both libchromium-efl.so and content_shell_efl
binaries. In ideal world content_shell_efl should only be composed of
sources from src/content/shell directory. Unfortunately upstream gyp
files add dependencies on various content components which are of
static_library type. The same components are also used by chromium-efl
target. The dynamic linker can't possibly figure out that some of those
static libraries are alrady part of libchromium-efl shared library and
thus it can't eliminate them. As a result the content_shell_efl is
slightly bigger than it should actually be.

For m34 this did not seem to be a problem. In m40 however we've observed
crashes caused by GetContentClient() function returning NULL value.
This was especially strange considering the fact the function bsically
returns a value of a global variable g_client. We always initialize it
to non NULL value early in the application startup and we never clear it
during application lifetime. The problem is that the variable is defined
in a source file which is a part of content_common build target which
both content_shell_lib and chromium-efl targets depend on. The
SetContentClient and GetContentClient functions are not considered
public and are not marked with CONTENT_EXPORT. As a result they're
hidden inside chromium-efl binary. Whenever the getter function will be
invoked from any code present in the chromium-efl binary it'll refer to
it's own 'copy' of the global variable. Code from content_shell_efl
will have 'copy' of the same global variable. As a result during startup
we initialized the version present in the content_shell_efl binary, but
never touched the one from libchromium-efl. When later GetContentClient
happens to be called from code being part of libchromium-efl.so it'll
return the value of the second, unitialized g_client.

Since fixing those problems at their core would require some bigger
changes to content_shell gyp/gypi files this patch uses an alternative,
but also effective solution. It changes the type of chromium-efl target
from shared to static library in case of content_shell builds. This
means the linker will only get object files when producing
content_shell_efl binary and will be able to remove duplicated code
parts. It's very similar to the upstream solution where content_shell is
always one fat binary that is invoked for all process types.

The downside is longer build times when content_shell application is
enabled. Also the --cotent-shell flag should not be used when
making normal builds which are supposed to be used for purpose of
running chromium layout tests.

With this patch in place the amount of layout test failures on beta
branch goes down from around 12k to only 1701.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=10219
Reviewed by: Antonio Gomes, Min-Soo Koo

Change-Id: Ia68841ce85bb3084c0376bce35b7ea95d1ca35ae
Signed-off-by: Piotr Tworek <p.tworek@samsung.com>