Implementation of ewk_notification_policies_removed.
This API function is for removing entries from the internal
engine's cache, while the real storage of the policies is
done on the client side (such as 'browser'). I've implemented
the cache (std::map) and added implementation of
webkitNotifications.checkPermission() to use this cache to check
for the permission (this was not implemented at all and
default implementation in blink returns always 'Allowed').
As the permission checking is done in the IO thread I've
made the NotificationControllerEfl a RefCountedThreadSafe
object, so it can be passed with the ResourceContext and
accessed without webview. To make this permissions cache
feature complete we need a function to set the initial state
(which e.g. browser could read from its database) - for
this purpose there is ewk_notification_cached_permissions_set()
function and this commit contains fuctionality needed by this
function.
EWK API regarding Notification's functionality is updated
in this patch. The "notification,permission,request" smart
callback is removed and replaced with a user defined callback
set with ewk_view_notification_permission_callback_set().
Other functions added in this patch:
  ewk_notification_permission_reply
Following functions are now deprecated:
  ewk_notification_permission_request_set
  ewk_notification_permission_request_suspend

This patch is a merge of the following patches:
http://165.213.202.130:8080/#/c/69294/
http://165.213.202.130:8080/#/c/70194/
http://165.213.202.130:8080/#/c/72200/
by Filip Piechocki

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=9559
Reviewed by: Antonio Gomes, Kamil Klimek, Piotr Tworek

Change-Id: I34c818a7f09735bcbf4aec4298851a2002e7bbc1
Signed-off-by: Tomasz Czekala <t.czekala@partner.samsung.com>