Reset pointer to BackForwardList on renderer crash
When renderer process will crash, new WebContents object
is created in order to open a page which informs that process
has crashed. WebContents object has NavigationControllerImpl
fiel. After creation of a new WebContents, old one is deleted
and along with it NavigationControllerImpl is. Thus any action
on it causes segmentation fault. We need to simply reset
NavigationController object. Also we need to reset
whole BackForwardList object, because we need to register
navigation notifications in a new context.

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=10304
Reviewed by: Kamil Klimek, Piotr Tworek, SeungSeop Park

Change-Id: I708b30c6db8c74284e30f078b8b04f39dff084cd
Signed-off-by: Daniel Waślicki <d.waslicki@samsung.com>